pyproject.toml
[tool.poetry]
name = "yankee"
version = "0.1.2"
description = "lightweight, simple, and fast declarative XML and JSON data extraction"
authors = ["Parker Hancock <633163+parkerhancock@users.noreply.github.com>"]
license = "Apache Software License 2.0"
readme = "README.md"
repository = "https://github.com/parkerhancock/gelatin_extract"
keywords = ["deserialization", "xml", "json", "deserialize"]
classifiers =[
    "Development Status :: 3 - Alpha",
    "Topic :: Text Processing :: Markup :: XML"
]

[tool.poetry.dependencies]
python = "^3.9"
lxml = "^4.8.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
keyring = "^23.5.1"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
